Refactor/Move login logic into services
