+++ server/graphql/resolvers/login/index.js
@@ -1,24 +1,5 @@
-import jwt from 'jsonwebtoken';
-import bcrypt from 'bcryptjs';
-import CustomError from '../../../customError';
-import User from '../../../models/user';
-
-const secret = process.env.SECRET;
+import getUser from '../../../services/getUser';
 
 export default async ({ email, password }) => {
-    try {
-        const user = await User.findOne({ email });
-        if (!user) {
-            throw new CustomError('User is not found', 'email', 'AuthUserError');
-        }
-        const isMatch = await bcrypt.compare(password, user.password);
-        if (isMatch) {
-            const token = await jwt.sign({ ...user._doc, password: null }, secret);
-            return { ...user._doc, password: null, token };
-        }
-        throw new CustomError('Password is incorrect', 'password', 'AuthUserError');
-    }
-    catch (e) {
-        return e;
-    }
+    return getUser({ email, password });
 };
