+++ server/graphql/resolvers/signup/index.js
@@ -1,4 +1,4 @@
-import postUser from '../../../postUser';
+import postUser from '../../../services/postUser';
 
 export default ({ userInput: { username, email, password } }) => {
     return postUser({ username, email, password });
